fix(server): use child LiveView for docs selector to fix event handling
The docs layout is a dead view (no LiveView socket), so NooraDropdown
hooks can't push events. Fix by extracting the selector into a
separate DocsSelectorLive child LiveView embedded in the layout via
live_render. This gives the selector its own LiveView connection for
handling select-account and select-project events.
