+++ server/lib/tuist_web/docs/layouts.ex
@@ -3,8 +3,6 @@ defmodule TuistWeb.Docs.Layouts do
   use TuistWeb, :html
   use Noora
 
-  import TuistWeb.AppLayoutComponents, only: [headerbar_breadcrumbs: 1]
-
   alias Tuist.Docs.Paths
 
   embed_templates "layouts/*"
