+++ server/lib/tuist_web/docs/layouts/root.html.heex
@@ -77,11 +77,12 @@
           <span data-part="logo-divider"></span>
           <span data-part="logo-text">Docs</span>
         </.link>
-        <.headerbar_breadcrumbs
-          :if={not is_nil(assigns[:current_user]) and (assigns[:breadcrumbs] || []) != []}
-          breadcrumbs={assigns[:breadcrumbs] || []}
-          id="docs-context-breadcrumbs"
-        />
+        <%= if not is_nil(assigns[:current_user]) do %>
+          {live_render(@conn, TuistWeb.DocsSelectorLive,
+            id: "docs-selector",
+            session: %{"current_user_id" => assigns[:current_user].id}
+          )}
+        <% end %>
         <.button_group size="large" id="docs-nav-tabs">
           <.button_group_item
             label={dgettext("docs", "Guides")}
